+++ CHANGES.md
@@ -3,6 +3,23 @@
 See [Microsoft's Change log](https://github.com/microsoft/SEAL/blob/master/Changes.md)
 for more details on each SEAL version change.
 
+## Version 3.1.0
+
+Fix:
+- `Context.getContextData({ parmsId })` now returns an instance of `ContextData`
+- `Context.keyContextData` now returns an instance of `ContextData`
+- `Context.firstContextData` now returns an instance of `ContextData`
+- `Context.lastContextData` now returns an instance of `ContextData`
+- `EncryptionParameters.plainModulus` now returns a wrapped instance of `SmallModulus`
+
+Feat:
+- Added `ContextData` binding which can be used for inspecting a `Context` information in greater detail.
+- Added `EncryptionParameterQualifiers` binding to be used for advanced debugging.
+
+Breaking:
+- `SmallModulus.value` now returns a BigInt containing the value stored inside 
+  the instance instead of a string.
+
 ## Version 3.0.6
 
 Fix: